% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbbga23}
\alias{cbbga23}
\title{Cbbga23 Data Set}
\format{
A data frame with 6 variables: \code{Date}, \code{Away}, \code{AwayPts},
\code{Home}, \code{HomePts}, \code{Diff}
}
\usage{
cbbga23
}
\description{
This data set contains information on college basketball games in 2023.
}
\keyword{datasets}
