% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamGms.R
\name{teamGms}
\alias{teamGms}
\title{A function to filter the cbbga23 data by games only played by a certain team}
\usage{
teamGms(data, team)
}
\arguments{
\item{data:}{the data to refer to (cbbga23)}

\item{team:}{The team to filter by, in quotation marks}
}
\value{
a tibble of games played by said team
}
\description{
A function to filter the cbbga23 data by games only played by a certain team
}
\examples{
#filter for only games played by Southern Utah University
teamGms(cbbga23, "Southern Utah")

}
