% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Teams}
\alias{Teams}
\title{Teams Data Set}
\format{
A data frame with 1 variables: \code{Team}
}
\usage{
Teams
}
\description{
This data set is a list of all teams who played at least one game in 2023
}
\keyword{datasets}
