% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmRec.R
\name{tmRec}
\alias{tmRec}
\title{A function to calculate the win/loss record of a given team}
\usage{
tmRec(team)
}
\arguments{
\item{x:}{The team to calculate record of}
}
\value{
the win/loss record of the team, as
}
\description{
A function to calculate the win/loss record of a given team
}
\examples{
#Calculate Southern Utah University's win loss record for the year of 2023
tmRec("Southern Utah")

}
