% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winLoss.R
\name{winLoss}
\alias{winLoss}
\title{A function to create a bar chart of team wins and losses}
\usage{
winLoss(team, colors)
}
\arguments{
\item{x}{: The team to plot, in quotation marks}

\item{colors:}{Colors for wins and losses, respectively}
}
\value{
a bar chart with a column for wins and losses
}
\description{
A function to create a bar chart of team wins and losses
}
\examples{
# Plot SUU's games, divided into wins and losses
winLoss("Southern Utah", colors = c("black", "red"))

}
